+++ sandbox/rocky/tf/spaces/box.py
@@ -7,4 +7,4 @@ import tensorflow as tf
 
 class Box(TheanoBox):
     def new_tensor_variable(self, name, extra_dims):
-        return tf.placeholder(tf.float32, shape=[None] * extra_dims + [self.flat_dim], name=name)
+        return tf.placeholder(tf.float64, shape=[None] * extra_dims + [self.flat_dim], name=name)
